refactor: move FormalDeMuxTester
